Turn talons and talons.auth into namspace pkgs
In order to use the same namespace "talons.auth" in external distrbutions
like talons.auth.oauth, it was necessary to turn talons and talons.auth
into namespace packages. Luckily, nobody has forked us so we're not in
danger of breaking any compatibility.
