+++ setup.py
@@ -18,4 +18,6 @@ import setuptools
 
 setuptools.setup(
     setup_requires=['d2to1>=0.2.10,<0.3', 'pbr>=0.5,<0.6'],
-    d2to1=True)
+    d2to1=True,
+    namespace_packages=['talons', 'talons.auth'],
+)
