+++ tox.ini
@@ -23,7 +23,7 @@ deps = pep8==1.4.5
        flake8==2.0
 commands =
     flake8
-    flake8 --filename=procession-*
+    flake8 --filename=talons-*
 
 [flake8]
 ignore = H301,H306
